feat (registrar/clients/websocket): add websocker connector with retry mechanism
